address review feedback: relocate EncoderGroup to mixin, tighten docstrings
- Move `EncoderGroup`, `_lengths_by_modality`, `_reorder_embeds_by_manifest`,
  and `_synthesize_single_modality_manifest` from `modeling_multimodal_utils.py`
  to `modeling_multimodal_mixin.py`; the framework is the only consumer.
- Freeze `EncoderGroup` (`@dataclass(frozen=True)`) and add per-field
  docstrings covering the `build_batched_input` / `encoder_fn` contract.
- Switch `:class:` / `:attr:` RST roles to single backticks in comments.
- Drop the `if total > 0` special case in `encode_multimodal_by_groups`;
  reuse `lengths` directly via `per_modality_lengths.update`.
- Replace the manual prefix-sum with `itertools.accumulate`.
- Add a comment in `_lengths_by_modality` explaining that raw-prompt
  entrypoints are the single enforcement point for the mm_item_order
  requirement on >1-modality requests.
- Update import sites in `modeling_qwen3vl.py` and the two multimodal tests.
